agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent"
dynamic = ["version"]
description = "Net load evolution with DERs"
license = "BSD-3-Clause"
requires-python = ">=3.10"
authors = [
    { name = "Kapil Duwadi", email = "kapil.duwadi@nrel.gov" },
    {name = "Erik Pohl", email = "erik.pohl@nrel.gov"}
]
keywords = [
    "Distribution",
    "Electric",
    "Energy",
    "Net",
    "PV",
    "Storage",
    "System",
    "Vehicle",
    "load",
]
classifiers = [
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    "polars",
    "pvlib",
    "pydantic",
    "pytest",
    "psycopg2-binary",
    "pika",
    "numpy",
    "python-dotenv",
    "pyarrow"
]

[project.urls]
Homepage = "https://github.com/NREL/evolve"

[tool.hatch.version]
path = "agent/__init__.py"

[tool.hatch.build.targets.sdist]
include = [
    "/agent",
]
